Create downloadCsvFile util
