+++ src/utils/csv.ts
@@ -1,0 +1,13 @@
+export function downloadCsvFile(filename: string, rows: string[][]) {
+  const csvContent =
+    'data:text/csv;charset=utf-8,' + rows.map(e => e.join(',')).join('\n')
+  const encodedUri = encodeURI(csvContent)
+
+  const link = document.createElement('a')
+  link.setAttribute('href', encodedUri)
+  link.setAttribute('download', `${filename}.csv`)
+
+  document.body.appendChild(link)
+
+  link.click()
+}
